stash: Restored `stash save --include-untracked` option support
The `stash--helper create --include-untracked` option was lost during the transition of git stash from shell to builtin.

https://github.com/git/git/commit/40af14683432285b94407e8488eab6942d0779dc#r153479277

I'm bringing it back.

This option is needed for non-destructive stashing of the the full state of the working directory.

Implementation details:
Unlike the `stash--helper.c` version, the `stash.c` version of `create_stash` treats all arguments as commit message. This makes adding/restoring options impossible without a breaking change. However we can heuristically detect options by checking whether the `argv` starts with a dash. I think this is a good compromise given the non-extensible design of the command. The only other alternative is to create an entirely new command (e.g. `stash save2`).